% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mywx-package.R
\docType{package}
\name{mywx}
\alias{mywx}
\alias{mywx-package}
\title{Tools to Query the 'MetMalaysia' Web Service 'API'}
\description{
Tools to Query the 'MetMalaysia' Web Service 'API'
}
\author{
Bob Rudis (bob@rud.is)
}
