% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast.r
\name{mywx_forecast}
\alias{mywx_forecast}
\title{Retrieve forecast data for a location}
\usage{
mywx_forecast(location_id, start_date = Sys.Date(), end_date = Sys.Date(),
  dataset_id = "FORECAST", data_category_id = "GENERAL",
  mywx_api_token = Sys.getenv("MYWX_TOKEN"))
}
\arguments{
\item{location_id}{valid location id retrieved from one of the location helpers}

\item{start_date, end_date}{either character vectors in ISO 8601 format or \code{Date}-like
objects. They both default to "today" if not specified}

\item{dataset_id}{the current only valid default, \code{FORECAST}, is pre-populated}

\item{data_category_id}{the current only valid default, \code{GENERAL} is pre-populated}

\item{mywx_api_token}{MET API token}
}
\description{
This is a WIP. The data returned is in a sub-optimal format that requires conversion of
some data frame column values to numeric to use.
}
\details{
You need a valid token either passed in to \code{mywx_api_token} or stored in the \code{MYWX_TOKEN},
preferably in \code{~/.Renviron}.
}
\examples{
\dontrun{
mywx_forecast("LOCATION:237", "2017-08-13", "2017-08-13")
mywx_forecast("LOCATION:237", "2017-08-01", "2017-08-13")
}
}
