% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/location.r
\name{mywx_states}
\alias{mywx_states}
\alias{mywx_districts}
\alias{mywx_towns}
\alias{mywx_touristdests}
\title{Retrieve valid geopolitical entities (location)}
\usage{
mywx_states(mywx_api_token = Sys.getenv("MYWX_TOKEN"))

mywx_districts(mywx_api_token = Sys.getenv("MYWX_TOKEN"))

mywx_towns(mywx_api_token = Sys.getenv("MYWX_TOKEN"))

mywx_touristdests(mywx_api_token = Sys.getenv("MYWX_TOKEN"))
}
\arguments{
\item{mywx_api_token}{MET API token}
}
\description{
Endpoint data relies on passing a \code{locationid}. These functions can be uses to obtain
a lookup table of valid loations and ids.
}
\details{
You need a valid token either passed in to \code{mywx_api_token} or stored in the \code{MYWX_TOKEN}
environment variable, preferably in \code{~/.Renviron}.
}
\examples{
\dontrun{
mywx_states()
mywx_districts()
mywx_towns()
mywx_touristdests()
}
}
